Refactor: LottoContainerFormatter 패키지 위치 조정
