+++ src/main/java/step2/domain/lotto/LottoContainer.java
@@ -1,6 +1,7 @@
 package step2.domain.lotto;
 
 import step2.domain.rank.Rank;
+import views.LottoContainerFormatter;
 import views.LottoFormatter;
 
 import java.util.ArrayList;
